requirements.txt
singer-python==5.9.0
singer-tools==0.4.1
requests==2.25.1
coverage